[Orders] Finished CRUD and model validation

+++ Orders/src/main/java/nwt/orders/repository/ReceiptsRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import nwt.orders.model.Receipt;
 
-public interface ReceiptsRepository extends JpaRepository<Receipt, Integer> {
+public interface ReceiptsRepository extends JpaRepository<Receipt, Long> {
 
 }
