+++ Orders/src/main/java/nwt/orders/repository/RentalsRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import nwt.orders.model.Rental;
 
 
-public interface RentalsRepository extends JpaRepository<Rental, Integer> {
+public interface RentalsRepository extends JpaRepository<Rental, Long> {
 
 }
