+++ Orders/src/main/java/nwt/orders/repository/ReturnsRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import nwt.orders.model.Returns;
 
-public interface ReturnsRepository extends JpaRepository<Returns, Integer> {
+public interface ReturnsRepository extends JpaRepository<Returns, Long> {
 
 }
